processing/cloud_function/D009/vectorize/requirements.txt
pymongo==4.8.0
Flask==2.1.1
Werkzeug==2.1.1
pydantic==2.9.2
python-magic==0.4.27
pandas==2.2.2
google-cloud-run==0.10.14